check-dns-soa-sync: major change in response validation
from now on the result is based only on the serial verification. Until now
it was based on the full SOA line, but the only required consistency is the
serial match. Most domains have the full SOA line consistency, though a small
percentage of the tested domains keep only the serial equal on all servers.
This change was made for this plugin to be more generic.